Reorganize previous sales work
