--- src/modules/dashboard/components/CoinMachine/PreviousSales/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PreviousSales';
