+++ src/modules/dashboard/components/CoinMachine/PreviousSalesCard.css
