+++ src/modules/dashboard/components/CoinMachine/PreviousSalesCard.css.d.ts
